change conrurrent to accept u64 instead of usize

+++ src/concurrent.rs
@@ -5,28 +5,28 @@ use std::future::Future;
 use std::sync::{Arc, Mutex};
 
 pub async fn run_concurrent<'a, F, Fut, V, E, M>(
-    start: usize,
-    end: usize,
+    start: u64,
+    end: u64,
     step: usize,
     middleware: Arc<M>,
     func: F,
 ) -> Result<Vec<V>, E>
 where
-    F: Fn(usize, usize, Arc<M>, Option<Arc<Mutex<ProgressBar>>>) -> Fut,
+    F: Fn(u64, u64, Arc<M>, Option<Arc<Mutex<ProgressBar>>>) -> Fut,
     Fut: Future<Output = Result<Vec<V>, E>> + Send + 'a,
     V: Send + 'a,
     E: Send + 'a,
     M: Middleware + 'a,
 {
     let size = end - start;
-    let pb = ProgressBar::new(size as u64);
+    let pb = ProgressBar::new(size);
     let shared_pb = Arc::new(Mutex::new(pb));
     let mut futures: Vec<_> = vec![];
 
     for i in (start..end).step_by(step) {
         futures.push(func(
             i,
-            (i + step).min(end),
+            (i + step as u64).min(end),
             middleware.clone(),
             Some(shared_pb.clone()),
         ));
